Update: 앨범 업로드 연결

+++ app/src/main/java/com/easyhz/picly/data/entity/album/ImageSize.kt
@@ -1,6 +1,6 @@
 package com.easyhz.picly.data.entity.album
 
 data class ImageSize(
-    val height: Int = 0,
-    val width: Int = 0
+    val height: Long = 0,
+    val width: Long = 0
 )
